requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
httpx>=0.27.0
google-genai>=0.4.0
python-dotenv>=1.0.1
